refactor: uniformize options API
Introduces InferenceOptions for inference scripts.

The options can also be used by the server and be parsed either as json
or command line.

Options are also better organized.

New tests for inference GAN + diffusion
