+++ models/base_diffusion_model.py
@@ -14,6 +14,7 @@ from tqdm import tqdm
 
 from util.diff_aug import DiffAugment
 from util.util import save_image, tensor2im
+from util.util import pairs_of_floats, pairs_of_ints, MAX_INT
 
 from .base_model import BaseModel
 from .modules.sam.sam_inference import (
