+++ models/cycle_gan_model.py
@@ -59,6 +59,10 @@ class CycleGanModel(BaseGanModel):
             )
         return parser
 
+    @staticmethod
+    def after_parse(opt):
+        return opt
+
     def __init__(self, opt, rank):
 
         super().__init__(opt, rank)
