+++ scripts/gen_single_image.py
@@ -2,9 +2,11 @@ import sys
 import os
 import json
 
-sys.path.append("../")
+jg_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "../")
+sys.path.append(jg_dir)
 from models import gan_networks
 from options.train_options import TrainOptions
+from options.inference_gan_options import InferenceGANOptions
 import cv2
 import torch
 from torchvision import transforms
@@ -28,7 +30,7 @@ def load_model(modelpath, model_in_file, cpu, gpuid):
     opt = TrainOptions().parse_json(train_json, set_device=False)
     if opt.model_multimodal:
         opt.model_input_nc += opt.train_mm_nz
-    opt.jg_dir = "../"
+    opt.jg_dir = jg_dir
 
     if not cpu:
         device = torch.device("cuda:" + str(gpuid))
@@ -45,70 +47,58 @@ def load_model(modelpath, model_in_file, cpu, gpuid):
     return model, opt, device
 
 
-parser = argparse.ArgumentParser()
-parser.add_argument(
-    "--model-in-file", help="file path to generator model (.pth file)", required=True
-)
+def inference(args):
+    modelpath = args.model_in_file.replace(os.path.basename(args.model_in_file), "")
+    print("modelpath=%s" % modelpath)
 
-parser.add_argument("--img-in", help="image to transform", required=True)
-parser.add_argument("--img-out", help="transformed image", required=True)
-parser.add_argument(
-    "--img-width", type=int, help="image width, defaults to model crop size"
-)
-parser.add_argument(
-    "--img-height", type=int, help="image height, defaults to model crop size"
-)
-parser.add_argument("--cpu", action="store_true", help="whether to use CPU")
-parser.add_argument("--gpuid", type=int, default=0, help="which GPU to use")
-args = parser.parse_args()
-
-# loading model
-modelpath = args.model_in_file.replace(os.path.basename(args.model_in_file), "")
-print("modelpath=", modelpath)
-
-model, opt, device = load_model(
-    modelpath, os.path.basename(args.model_in_file), args.cpu, args.gpuid
-)
-
-# reading image
-img_width = args.img_width if args.img_width is not None else opt.data_crop_size
-img_height = args.img_height if args.img_height is not None else opt.data_crop_size
-img = cv2.imread(args.img_in)
-img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-img = cv2.resize(img, (img_width, img_height), interpolation=cv2.INTER_CUBIC)
-
-# preprocessing
-tranlist = [
-    transforms.ToTensor(),
-    transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5)),
-]
-tran = transforms.Compose(tranlist)
-img_tensor = tran(img)
-if not args.cpu:
-    img_tensor = img_tensor.to(device)
-
-if opt.model_multimodal:
-    z_random = get_z_random(batch_size=1, nz=opt.train_mm_nz)
-    z_random = z_random.to(device)
-    # print('z_random shape=', self.z_random.shape)
-    z_real = z_random.view(z_random.size(0), z_random.size(1), 1, 1).expand(
-        z_random.size(0),
-        z_random.size(1),
-        img_tensor.size(1),
-        img_tensor.size(2),
+    model, opt, device = load_model(
+        modelpath, os.path.basename(args.model_in_file), args.cpu, args.gpuid
     )
-    img_tensor = torch.cat([img_tensor.unsqueeze(0), z_real], 1)
-else:
-    img_tensor = img_tensor.unsqueeze(0)
 
-# run through model
-out_tensor = model(img_tensor)[0].detach()
+    # reading image
+    img_width = args.img_width if args.img_width is not None else opt.data_crop_size
+    img_height = args.img_height if args.img_height is not None else opt.data_crop_size
+    img = cv2.imread(args.img_in)
+    img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+    img = cv2.resize(img, (img_width, img_height), interpolation=cv2.INTER_CUBIC)
 
-# post-processing
-out_img = out_tensor.data.cpu().float().numpy()
-print(out_img.shape)
-out_img = (np.transpose(out_img, (1, 2, 0)) + 1) / 2.0 * 255.0
-# print(out_img)
-out_img = cv2.cvtColor(out_img, cv2.COLOR_RGB2BGR)
-cv2.imwrite(args.img_out, out_img)
-print("Successfully generated image ", args.img_out)
+    # preprocessing
+    tranlist = [
+        transforms.ToTensor(),
+        transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5)),
+    ]
+    tran = transforms.Compose(tranlist)
+    img_tensor = tran(img)
+    if not args.cpu:
+        img_tensor = img_tensor.to(device)
+
+    if opt.model_multimodal:
+        z_random = get_z_random(batch_size=1, nz=opt.train_mm_nz)
+        z_random = z_random.to(device)
+        # print('z_random shape=', self.z_random.shape)
+        z_real = z_random.view(z_random.size(0), z_random.size(1), 1, 1).expand(
+            z_random.size(0),
+            z_random.size(1),
+            img_tensor.size(1),
+            img_tensor.size(2),
+        )
+        img_tensor = torch.cat([img_tensor.unsqueeze(0), z_real], 1)
+    else:
+        img_tensor = img_tensor.unsqueeze(0)
+
+    # run through model
+    out_tensor = model(img_tensor)[0].detach()
+
+    # post-processing
+    out_img = out_tensor.data.cpu().float().numpy()
+    print(out_img.shape)
+    out_img = (np.transpose(out_img, (1, 2, 0)) + 1) / 2.0 * 255.0
+    # print(out_img)
+    out_img = cv2.cvtColor(out_img, cv2.COLOR_RGB2BGR)
+    cv2.imwrite(args.img_out, out_img)
+    print("Successfully generated image ", args.img_out)
+
+
+if __name__ == "__main__":
+    opt = InferenceGANOptions().parse()
+    inference(opt)
